requirements.txt
Django==4.2.11
django-webtest==1.9.11
djangorestframework==3.15.0
PyHamcrest==2.1.0
WebTest==3.0.0